refactor(benchmark): extract clap CLI types into library module
Move `Cli`, `Commands`, the seven subcommand enums, `Tool`, `GenomeOption`,
`parse_genome_option`, and `CompareMode` from `src/bin/benchmark.rs` into a
new `pub mod ferro_hgvs::benchmark::cli`. The binary becomes a thin shim
that imports them and delegates to the same handler bodies.

This keeps the parser definition reachable from integration tests, so they
can call `Cli::try_parse_from(...)` in-process instead of spawning the
binary via `cargo run --release`.

No behavioral change to the CLI surface.
